# Increased accuracy when getting current audio time using OpenAL SOFT on native