--- a/src/lime/media/OpenALAudioContext.hx
+++ b/src/lime/media/OpenALAudioContext.hx
@@ -91,6 +91,13 @@ class OpenALAudioContext
 	public var DEFAULT_ALL_DEVICES_SPECIFIER:Int = 0x1012;
 	public var ALL_DEVICES_SPECIFIER:Int = 0x1013;
 
+	public var DEVICE_CLOCK_SOFT:Int = 0x1600;
+	public var DEVICE_LATENCY_SOFT:Int = 0x1601;
+	public var DEVICE_CLOCK_LATENCY_SOFT:Int = 0x1602;
+
+	public var SEC_OFFSET_LATENCY_SOFT:Int = 0x1201;
+	public var SEC_OFFSET_CLOCK_SOFT:Int = 0x1203;
+
 	@:noCompletion private function new() {}
 
 	public function bufferData(buffer:ALBuffer, format:Int, data:ArrayBufferView, size:Int, freq:Int):Void