--- a/src/lime/media/openal/AL.hx
+++ b/src/lime/media/openal/AL.hx
@@ -977,7 +977,7 @@ class AL
 	public static function getSourcedvSOFT(source:ALSource, param:Int, count:Int = 1):Array<Float>
 	{
 		#if (lime_cffi && lime_openal && !macro)
-		var result = NativeCFFI.lime_al_get_sourcedv_SOFT(source, param, count);
+		var result = NativeCFFI.lime_al_get_sourcedv_soft(source, param, count);
 		#if hl
 		if (result == null) return [];
 		var _result:Array<Float> = [];