@ninjamuffin99 2024-02-09
cleaned up the whitespacing, and removed those unneeded public vars 🤝