@SomeGuyWhoLovesCoding 2024-05-19
> openal soft gives null object referance at iOS ![IMG_3256](https://private-user-images.githubusercontent.com/82843871/331859349-7aca5ec6-4c87-406f-a5bb-6739490f0593.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTYxMjcwNzMsIm5iZiI6MTcxNjEyNjc3MywicGF0aCI6Ii84Mjg0Mzg3MS8zMzE4NTkzNDktN2FjYTVlYzYtNGM4Ny00MDZmLWE1YmItNjczOTQ5MGYwNTkzLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA1MTklMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNTE5VDEzNTI1M1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTMyNWRlYTZmODFkM2YwZTdiMmRlY2IzOGQ4NjM0M2JlMWM3Yjk0ZWY4YzY5M2M5ZjM5NjQ2MjlmNjhlNDYxZjgmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.tfONz9dX6Brv9yBWZvgo-oLz0CgngnZGbA9ekSprZek)

Ikr? LMAO 🤣