@mcagabe19 2024-05-30
> Maybe a define needs to be added as openal soft is not on ios or lime should add support for openal soft on ios.

OpenAL SOFT is not supported for `static_link` or for iOS itself so lime needed add support for it