@joshtynjala 2024-05-30
> OpenAL SOFT is not supported for static_link or for iOS itself so lime needed add support for it

OpenAL SOFT is LGPL, so as I understand it, it must be dynamically linked.