@mcagabe19 2024-06-06
> OpenAL SOFT is LGPL, so as I understand it, it must be dynamically linked.

Makes sense, but still OpenAL SOFT needs iOS fixes/patches