@ACrazyTown 2024-07-18
> > Any updates on this?
> 
> iOS is currently broken

Yeah I saw, I asked to see if there was any progress on fixing it and getting this ready to merge