@Dimensionscape 2024-07-19
> > > Any updates on this?
> > 
> > 
> > iOS is currently broken
> 
> Yeah I saw, I asked to see if there was any progress on fixing it and getting this ready to merge

We dont use OpenAL when static linking. Instead we use MojoAL. see: https://github.com/openfl/lime/blob/ca52db81efa10b1bd8c76e43c714faa9dbbc4c94/project/Build.xml#L26