@Dimensionscape 2024-07-19
We don't use OpenAL Soft when static linking. Instead we use MojoAL. 

See: https://github.com/openfl/lime/blob/ca52db81efa10b1bd8c76e43c714faa9dbbc4c94/project/Build.xml#L26

From what I understand, apple devices have the proprietary licensed OpenAL embedded in the system api's(as opposed to OpenAL Soft) and we use that for those targets [here](https://github.com/openfl/lime/blob/ca52db81efa10b1bd8c76e43c714faa9dbbc4c94/project/src/media/openal/OpenALBindings.cpp#L1C1-L3C24).

Any changes need to take account for our different OpenAL backends.