@Dimensionscape 2024-07-22
It feels a little dirty having a function that only works for a specific AL backend, but it sounds like it should be something we can manage uniformly. ✌