@ACrazyTown 2024-09-13
Just wondering, would [setCurrentTime](https://github.com/FunkinCrew/lime/blob/openal-soft-pr/src/lime/_internal/backend/native/NativeAudioSource.hx#L422) also have to account for these changes? I see it references `AL.BYTE_OFFSET` like the old audio time method did so I'm not quite sure